[Feat] #31 - API 종류 분리
